+++ chats/apps/api/v1/projects/serializers.py
@@ -1,7 +1,8 @@
 from rest_framework import serializers
 from timezone_field.rest_framework import TimeZoneSerializerField
 
-from chats.apps.projects.models import LinkContact, Project
+from chats.apps.projects.models import FlowStart, LinkContact, Project
+from chats.apps.sectors.models import Sector
 
 
 class ProjectSerializer(serializers.ModelSerializer):
@@ -61,3 +62,24 @@ class ProjectFlowStartSerializer(serializers.Serializer):
         trim_whitespace=True,
     )
     params = serializers.JSONField(required=False)
+    contact_name = serializers.CharField()
+
+
+class ListFlowStartSerializer(serializers.ModelSerializer):
+    user = serializers.SerializerMethodField()
+
+    class Meta:
+        model = FlowStart
+        fields = ["contact_data", "name", "user", "created_on", "room"]
+
+    def get_user(self, flow_start: FlowStart) -> str:
+        try:
+            return flow_start.permission.user.full_name
+        except AttributeError:
+            return ""
+
+
+class SectorDiscussionSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Sector
+        fields = ["uuid", "name"]