+++ chats/apps/projects/admin.py
@@ -1,10 +1,11 @@
 from django.contrib import admin
 
 from chats.apps.projects.models import (
+    FlowStart,
+    LinkContact,
     Project,
     ProjectPermission,
     TemplateType,
-    LinkContact,
 )
 
 # Register your models here.
@@ -13,3 +14,4 @@ admin.site.register(Project)
 admin.site.register(ProjectPermission)
 admin.site.register(TemplateType)
 admin.site.register(LinkContact)
+admin.site.register(FlowStart)